Move selector opts to just manual reader
