+++ sdk/metric/reader.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"fmt"
 
-	"go.opentelemetry.io/otel/internal/global"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/metricdata"
 	"go.opentelemetry.io/otel/sdk/metric/view"
@@ -108,13 +107,6 @@ func (p shutdownProducer) produce(context.Context) (metricdata.ResourceMetrics, 
 	return metricdata.ResourceMetrics{}, ErrReaderShutdown
 }
 
-// ReaderOption applies a configuration option value to either a ManualReader or
-// a PeriodicReader.
-type ReaderOption interface {
-	ManualReaderOption
-	PeriodicReaderOption
-}
-
 // TemporalitySelector selects the temporality to use based on the InstrumentKind.
 type TemporalitySelector func(view.InstrumentKind) metricdata.Temporality
 
@@ -123,29 +115,6 @@ type TemporalitySelector func(view.InstrumentKind) metricdata.Temporality
 // for all instrument kinds if this TemporalitySelector is used.
 func DefaultTemporalitySelector(view.InstrumentKind) metricdata.Temporality {
 	return metricdata.CumulativeTemporality
-}
-
-// WithTemporalitySelector sets the TemporalitySelector a reader will use to
-// determine the Temporality of an instrument based on its kind. If this
-// option is not used, the reader will use the DefaultTemporalitySelector.
-func WithTemporalitySelector(selector TemporalitySelector) ReaderOption {
-	return temporalitySelectorOption{selector: selector}
-}
-
-type temporalitySelectorOption struct {
-	selector func(instrument view.InstrumentKind) metricdata.Temporality
-}
-
-// applyManual returns a manualReaderConfig with option applied.
-func (t temporalitySelectorOption) applyManual(mrc manualReaderConfig) manualReaderConfig {
-	mrc.temporalitySelector = t.selector
-	return mrc
-}
-
-// applyPeriodic returns a periodicReaderConfig with option applied.
-func (t temporalitySelectorOption) applyPeriodic(prc periodicReaderConfig) periodicReaderConfig {
-	prc.temporalitySelector = t.selector
-	return prc
 }
 
 // AggregationSelector selects the aggregation and the parameters to use for
@@ -171,43 +140,4 @@ func DefaultAggregationSelector(ik view.InstrumentKind) aggregation.Aggregation 
 		}
 	}
 	panic("unknown instrument kind")
-}
-
-// WithAggregationSelector sets the AggregationSelector a reader will use to
-// determine the aggregation to use for an instrument based on its kind. If
-// this option is not used, the reader will use the DefaultAggregationSelector
-// or the aggregation explicitly passed for a view matching an instrument.
-func WithAggregationSelector(selector AggregationSelector) ReaderOption {
-	// Deep copy and validate before using.
-	wrapped := func(ik view.InstrumentKind) aggregation.Aggregation {
-		a := selector(ik)
-		cpA := a.Copy()
-		if err := cpA.Err(); err != nil {
-			cpA = DefaultAggregationSelector(ik)
-			global.Error(
-				err, "using default aggregation instead",
-				"aggregation", a,
-				"replacement", cpA,
-			)
-		}
-		return cpA
-	}
-
-	return aggregationSelectorOption{selector: wrapped}
-}
-
-type aggregationSelectorOption struct {
-	selector AggregationSelector
-}
-
-// applyManual returns a manualReaderConfig with option applied.
-func (t aggregationSelectorOption) applyManual(c manualReaderConfig) manualReaderConfig {
-	c.aggregationSelector = t.selector
-	return c
-}
-
-// applyPeriodic returns a periodicReaderConfig with option applied.
-func (t aggregationSelectorOption) applyPeriodic(c periodicReaderConfig) periodicReaderConfig {
-	c.aggregationSelector = t.selector
-	return c
 }
